backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
strawberry-graphql[fastapi]==0.243.1
python-dotenv==1.0.1
pydantic>=2.0,<2.11
psycopg[binary]==3.2.3
psycopg-pool==3.2.3
